requirements.txt
Cython
javaproperties
argparse
pillow
pycocotools
